+++ spec/support/site/pages/account_page.rb
@@ -1,0 +1,3 @@
+class AccountPage < Recurlysite::Page
+    set_url "/accounts/{account_code}"
+end
