+++ spec/support/site/pages/accounts_page.rb
@@ -1,0 +1,3 @@
+class AccountsPage < Recurlysite::Page
+    set_url "/accounts"
+end
