+++ spec/support/site/site.rb
@@ -1,0 +1,3 @@
+class Recurlysite
+  include Insite
+end
